checklist validator wf
